configs: atmel: sama5d4_xplained_mmc: enable HW PMECC
NAND:  BUG at drivers/mtd/nand/raw/nand_base.c:4361/nand_scan_tail()!

This board has a NAND flash enabled, and it requires the HW PMECC in order
to correctly probe this flash.
In the NAND flash configuration , this is selected by the GENERATE_PMECC_HEADER
which is not needed for mmc configuration.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>